+++ API/src/main/java/lib/brainsynder/nms/key/v8_to_v12/BaseParticlePacket.java
@@ -21,9 +21,9 @@ public class BaseParticlePacket extends ParticlePacket {
                 packetConstructor = packetClass.getDeclaredConstructor(
                         enumParticle,
                         Boolean.TYPE,
-                        Float.TYPE,
-                        Float.TYPE,
-                        Float.TYPE,
+                        Double.TYPE,
+                        Double.TYPE,
+                        Double.TYPE,
                         Float.TYPE,
                         Float.TYPE,
                         Float.TYPE,
